backend/requirements.txt
blinker==1.9.0
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
holidays==0.69
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
prophet==1.1.6
psycopg2-binary==2.9.9
pyparsing==3.2.3
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
stanio==0.5.1
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing==3.7.4.3
typing_extensions==4.13.1
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1